raceenabled works now
